pyproject.toml
[tool.poetry]
name = "dz-drf-courses"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
ipython = "^8.30.0"


[tool.poetry.group.lint.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
